refactor: photo_core を dating/exif_info/layout サブモジュールへ分割（振る舞い不変） (#12)
photo_core.rs(1097行) を photo_core/ ディレクトリモジュール化し責務分割:
- dating.rs: extract_date_from_filename / get_file_created_date / get_file_modified_date / format_filename（+characterization テスト22件）
- exif_info.rs: get_exif_info / is_image_file / is_video_file / ExifInfo（モジュール名は exif クレートとの衝突回避）
- layout.rs: create_date_hierarchy / create_backup / create_unsorted_dir
- mod.rs: 公開型（MediaInfo/ProcessOptions/ProcessResult/MediaType/DateSource 等）とパイプライン（scan_media/process_media/process_media_with_list）を維持

外部パス（photo_core::scan_media 等）は不変。lib.rs / cli.rs は無変更。
関数の中身・引数・戻り値は変えず純粋な移動＋可視性調整（pub(crate)）のみ。
cargo test 27件緑 / clippy 0 warning / fmt 適用。docs/development.md のモジュール記述を更新。
